feat: separação de cada funcionalidade do controllers em arquivos
