+++ src/bank/index.ts
@@ -1,25 +1,19 @@
-import {
-  registerBank,
-  loginBank,
-  searchMyBank,
-  getAllAccountsOfMyBank,
-  deleteBank,
-  updateDataBank,
-} from './controllers/controllers.banks';
-import {
-  midBankRegister,
-  midBankLogin,
-  midUpdateBank,
-} from './middlewares/middlewares.banks';
+import { deleteValidation, deleteBank } from './controllers/delete';
+import { registerValidation, registerBank } from './controllers/register';
+import { loginValidation, loginBank } from './controllers/login';
+import { searchMyBank } from './controllers/searchMyBank';
+import { getAllAccountsOfMyBank } from './controllers/getAllAccountsOfMyBank';
+import { updateValidation, update } from './controllers/update';
 
 export const bank = {
-  midBankRegister,
-  midBankLogin,
-  midUpdateBank,
+  registerValidation,
   registerBank,
+  loginValidation,
   loginBank,
   searchMyBank,
   getAllAccountsOfMyBank,
+  updateValidation,
+  update,
+  deleteValidation,
   deleteBank,
-  updateDataBank,
 };
